(editor) Share Translations & strings section with the badge editor
The locale override + custom-phrases section (and _resetPhrases) lived only in
the card editor, so the badge had no UI to change the display language or
customize allergen names. Move _renderPhrasesSection() + _resetPhrases() +
_selectedPhraseLang into the shared PollenEditorBase, and render it in the badge
editor too. The section is null-safe on config.phrases and gates its level-name
and day-label sub-blocks behind _showPhraseLevels() / _showPhraseDays() hooks;
the badge (which shows only the allergen name) overrides both to false, so it
exposes the language selector + full/short allergen names only.

The card editor now delegates to the shared method (behaviour unchanged) and
drops the imports the inline copy needed. Badge runtime already preserves
config.phrases / date_locale and the adapter localizes at fetch, so no badge
element change is needed. No new i18n keys.
